+++ src/main/java/de/maxhenkel/corpse/gui/TransferItemsButton.java
@@ -3,7 +3,7 @@ package de.maxhenkel.corpse.gui;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractButton;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
-import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.RenderPipelines;
 import net.minecraft.network.chat.Component;
 
 import java.util.function.Consumer;
@@ -23,9 +23,9 @@ public class TransferItemsButton extends AbstractButton {
     @Override
     public void renderWidget(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTicks) {
         if (isHovered) {
-            guiGraphics.blit(RenderType::guiTextured, CorpseAdditionalScreen.CORPSE_GUI_TEXTURE, getX(), getY(), 176, 10, WIDTH, HEIGHT, 256, 256);
+            guiGraphics.blit(RenderPipelines.GUI_TEXTURED, CorpseAdditionalScreen.CORPSE_GUI_TEXTURE, getX(), getY(), 176, 10, WIDTH, HEIGHT, 256, 256);
         } else {
-            guiGraphics.blit(RenderType::guiTextured, CorpseAdditionalScreen.CORPSE_GUI_TEXTURE, getX(), getY(), 176, 0, WIDTH, HEIGHT, 256, 256);
+            guiGraphics.blit(RenderPipelines.GUI_TEXTURED, CorpseAdditionalScreen.CORPSE_GUI_TEXTURE, getX(), getY(), 176, 0, WIDTH, HEIGHT, 256, 256);
         }
     }