+++ src/main/java/de/maxhenkel/corpse/net/MessageOpenHistory.java
@@ -51,11 +51,13 @@ public class MessageOpenHistory implements Message<MessageOpenHistory> {
     @Override
     public MessageOpenHistory fromBytes(RegistryFriendlyByteBuf buf) {
         CompoundTag compound = buf.readNbt();
-        ListTag list = compound.getList("Deaths", 10);
+        ListTag list = compound.getListOrEmpty("Deaths");
 
         deaths = new ArrayList<>();
         for (int i = 0; i < list.size(); i++) {
-            deaths.add(Death.fromNBT(buf.registryAccess(), list.getCompound(i)));
+            list.getCompound(i).ifPresent(e -> {
+                deaths.add(Death.read(buf.registryAccess(), e));
+            });
         }
 
         return this;
@@ -67,7 +69,7 @@ public class MessageOpenHistory implements Message<MessageOpenHistory> {
 
         ListTag list = new ListTag();
         for (Death d : deaths) {
-            CompoundTag c = d.toNBT(buf.registryAccess(), false);
+            CompoundTag c = d.write(buf.registryAccess(), false);
             list.add(c);
         }